Add build, install, and test macros for CMake
CMake supports multiple generator types for builds, such as GNU Make
and Ninja. As of CMake 3.15, there are now analogs for build and install
steps typically managed directly by build tools, where CMake controls
the logic to correctly call out.

Consequently, it is now rather easy to support any given type of
build tool (as long as it's available in the build environment) and
have a CMake build "work" as expected.

This also brings CMake in line with the same build/install process
that is used for packaging Meson-based projects.