Adjust ctest macro to use native out-of-tree support
With CMake 3.20, ctest(1) learned how to be told where the generated
working directory is, so we no longer have to change directories
(and thus, rely on shell semantics to get back out of that directory)
to make CTest work properly.

This change tweaks the %ctest macro such that we are leveraging that
new option and eliminates our usage of cd(1) as a workaround.