Only require cmake-rpm-macros when rpm-build is installed
There are other use-cases for cmake besides building rpms, for which a hard
requirement of rpm is absurd. Ironically enough, this hard dependency also
causes an ugly conflict requiring --nodeps hacks when building rpm's own
test-suite.

A conditional dependency on rpm-build sorts the issue nicely with no
regressions to the use-case of building packages.